requirements.txt
streamlit>=1.36
pandas>=2.0
python-dotenv>=1.0
